feat: add thunk for registration
